requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
requests>=2.25.1
python-dotenv>=0.19.0
grpcio>=1.40.0
grpcio-tools>=1.40.0
grpcio-reflection>=1.40.0
protobuf>=3.17.3